fm-ucode: new version 107_4_2_candidate1
Features: The following IP Acceleration features are not
supported in this package:
- IP fragmentation,
- IP reassembly,
- Header Manipulation,
- Frame Replication.
Please note that CC STD (statistics table descriptor) is supported.

Bug Fixes/CCB
This release is a result of integration with Linux environment and
full power down management system
testing. The following are the work-arounds implemented:
- Set FMFP_BRKC[SLP] bit just before event of wake-up.
- Eliminate internal syncbefore transmitting AR packets.

Known Issues
The following FMD updates to the deep sleep transition steps are:
- The Auto-Response TX IM port is enabled during deep sleep and
  its FMBM_TFDNE set to NIA 0x005000C0 (BMI Release).
- The FMFP_BRKC[SLP] is not set by the FMD instead it is set in firmware.

Signed-off-by: Mandy Lavi <mandy.lavi@freescale.com>
Change-Id: I7b3c6f87eacc4ed2137171ef43ec148f4a54f458
Reviewed-on: http://git.am.freescale.net:8181/13051
Reviewed-by: Mandy Lavi <Mandy.Lavi@freescale.com>
Tested-by: Mandy Lavi <Mandy.Lavi@freescale.com>